move subprocess to a util file
