+++ .github/workflows/actions/configuration/action.yaml
@@ -52,7 +52,7 @@ runs:
     # compatibility test depends on operator & apiserver images built in previous steps.
     run: |
       echo "Using Ray image ${{ inputs.ray_version }}"
-      cd tests/framework
+      cd tests
       GITHUB_ACTIONS=true \
       RAY_IMAGE="rayproject/ray:${{ inputs.ray_version }}" \
       OPERATOR_IMAGE="kuberay/operator:${{ steps.vars.outputs.sha_short }}" \
