+++ tests/framework/utils.py
@@ -1,0 +1,27 @@
+"""Utilities for integration tests of KubeRay."""
+
+import subprocess
+import logging
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(
+    format='%(asctime)s,%(msecs)d %(levelname)-8s [%(filename)s:%(lineno)d] %(message)s',
+    datefmt='%Y-%m-%d:%H:%M:%S',
+    level=logging.INFO)
+
+def shell_subprocess_run(command, check = True):
+    """
+    Command will be executed through the shell. If check=True, it will raise an error when
+    the returncode of the execution is not 0.
+    """
+    logger.info("Execute command: %s", command)
+    return subprocess.run(command, shell = True, check = check).returncode
+
+def shell_subprocess_check_output(command):
+    """
+    Run command and return STDOUT as encoded bytes.
+    """
+    logger.info("Execute command (check_output): %s", command)
+    output = subprocess.check_output(command, shell=True)
+    logger.info("Output: %s", output)
+    return output
