+++ tests/kuberay_utils/utils.py
@@ -13,11 +13,14 @@ from framework.prototype import (
     CONST,
     K8S_CLUSTER_MANAGER,
     RayClusterAddCREvent,
-    RayServiceAddCREvent,
+    RayServiceAddCREvent
+)
+
+from framework.utils import (
     shell_subprocess_run
 )
 
-raycluster_service_file = 'tests/config/raycluster-service.yaml'
+raycluster_service_file = CONST.REPO_ROOT.joinpath("tests/config/raycluster-service.yaml")
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
